app/requirements.txt
asgiref==3.2.10
boto3==1.14.49
botocore==1.17.49
dj-database-url==0.5.0
Django==3.1
django-crispy-forms==1.9.2
django-filter==2.3.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.2.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
